+++ crates/stackable-versioned-macros/src/gen/vstruct.rs
@@ -2,7 +2,7 @@ use darling::FromField;
 use itertools::Itertools;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
-use syn::{DataStruct, Error, Ident, Result};
+use syn::{Attribute, DataStruct, Error, Ident, Result};
 
 use crate::{
     attrs::{container::ContainerAttributes, field::FieldAttributes},
@@ -30,6 +30,9 @@ pub(crate) struct VersionedStruct {
     pub(crate) fields: Vec<VersionedField>,
 
     pub(crate) skip_from: bool,
+
+    /// The original attributes that were added to the struct.
+    pub(crate) original_attrs: Vec<Attribute>,
 }
 
 impl VersionedStruct {
@@ -37,6 +40,7 @@ impl VersionedStruct {
         ident: Ident,
         data: DataStruct,
         attributes: ContainerAttributes,
+        original_attrs: Vec<Attribute>,
     ) -> Result<Self> {
         // Convert the raw version attributes into a container version.
         let versions = attributes
@@ -46,6 +50,7 @@ impl VersionedStruct {
                 skip_from: v.skip.as_ref().map_or(false, |s| s.from.is_present()),
                 ident: format_ident!("{version}", version = v.name.to_string()),
                 deprecated: v.deprecated.is_present(),
+                doc: v.doc.clone(),
                 inner: v.name,
             })
             .collect();
@@ -98,6 +103,7 @@ impl VersionedStruct {
             versions,
             fields,
             ident,
+            original_attrs,
         })
     }
 
@@ -135,12 +141,24 @@ impl VersionedStruct {
 
         let deprecated_attr = version.deprecated.then_some(quote! {#[deprecated]});
         let module_name = &version.ident;
+        let attrs = &self.original_attrs;
+        let doc = if let Some(doc) = &version.doc {
+            let doc = format!("Docs for `{module_name}`: {doc}");
+            Some(quote! {
+                #[doc = ""]
+                #[doc = #doc]
+            })
+        } else {
+            None
+        };
 
         // Generate tokens for the module and the contained struct
         token_stream.extend(quote! {
             #[automatically_derived]
             #deprecated_attr
             pub mod #module_name {
+                #(#attrs)*
+                #doc
                 pub struct #struct_name {
                     #fields
                 }